backend/requirements.txt
fastapi
uvicorn[standard]
pymongo
pydantic[email]
pyjwt
python-dotenv
python-multipart